+++ R/derive_merged.R
@@ -721,10 +721,17 @@ derive_var_merged_exist_flag <- function(dataset,
 #'
 #' @permitted [dataset]
 #'
-#' @param print_not_mapped Print a list of unique `by_vars` values that do not
-#' have corresponding records from the lookup table?
+#' @param print_not_mapped `r lifecycle::badge("deprecated")` Print a list of unique
+#' `by_vars` values that do not have corresponding records from the lookup table?
 #'
 #' @permitted [boolean]
+#'
+#' @param check_not_mapped_type Check if any observations are not mapped?
+#'
+#'   If `"warning"`, `"message"`, or `"error"` is specified, the specified message is issued
+#'   if some records from the input dataset are not mapped using the lookup table.
+#'
+#' @permitted [msg_type]
 #'
 #' @inheritParams derive_vars_merged
 #'
@@ -793,10 +800,17 @@ derive_vars_merged_lookup <- function(dataset,
                                       filter_add = NULL,
                                       check_type = "warning",
                                       duplicate_msg = NULL,
-                                      print_not_mapped = TRUE) {
+                                      print_not_mapped = NULL,
+                                      check_not_mapped_type = "message") {
   by_vars_left <- replace_values_by_names(by_vars)
-  assert_logical_scalar(print_not_mapped)
+  assert_logical_scalar(print_not_mapped, optional = TRUE)
   filter_add <- assert_filter_cond(enexpr(filter_add), optional = TRUE)
+  check_not_mapped_type <-
+    assert_character_scalar(
+      check_not_mapped_type,
+      values = c("none", "warning", "error", "message"),
+      case_sensitive = FALSE
+    )
 
   tmp_lookup_flag <- get_new_tmp_var(dataset_add, prefix = "tmp_lookup_flag")
 
@@ -813,12 +827,16 @@ derive_vars_merged_lookup <- function(dataset,
     duplicate_msg = duplicate_msg
   )
 
-  if (print_not_mapped) {
+  if ((!is.null(print_not_mapped) && print_not_mapped) || check_not_mapped_type != "none") {
+    # Identify if any unmapped records exist
     temp_not_mapped <- res %>%
       filter(is.na(!!tmp_lookup_flag)) %>%
       distinct(!!!by_vars_left)
 
-    if (nrow(temp_not_mapped) > 0) {
+    some_not_mapped <- nrow(temp_not_mapped) > 0
+
+    # Store unmapped records
+    if (some_not_mapped) {
       # nolint start: undesirable_function_linter
       admiral_environment$nmap <- structure(
         temp_not_mapped,
@@ -826,14 +844,49 @@ derive_vars_merged_lookup <- function(dataset,
         by_vars = vars2chr(by_vars_left)
       )
       # nolint end
+    }
 
-      cli_inform(
+    if (!is.null(print_not_mapped)) {
+      deprecate_inform(
+        when = "1.6.0",
+        what = "derive_vars_merged_lookup(print_not_mapped)",
+        with = "derive_vars_merged_lookup(check_not_mapped_type)",
+        details = c(
+          x = "This message will turn into a warning at the beginning of 2028.",
+          # nolint start: line_length_linter
+          i = "See admiral's deprecation guidance:
+              https://pharmaverse.github.io/admiraldev/dev/articles/programming_strategy.html#deprecation"
+          # nolint end: line_length_linter
+        )
+      )
+      if (print_not_mapped && some_not_mapped) {
+        cli_inform(
+          c("List of {.var {vars2chr(by_vars_left)}} not mapped:",
+            capture.output(temp_not_mapped),
+            i = "Run {.run admiral::get_not_mapped()} to access the full list."
+          )
+        )
+      } else {
+        cli_inform(
+          "All {.var {vars2chr(by_vars_left)}} are mapped."
+        )
+      }
+    }
+
+    if (check_not_mapped_type != "none" && some_not_mapped) {
+      cli_function <- switch(check_not_mapped_type,
+        warning = cli_warn,
+        message = cli_inform,
+        error = cli_abort
+      )
+
+      cli_function(
         c("List of {.var {vars2chr(by_vars_left)}} not mapped:",
           capture.output(temp_not_mapped),
           i = "Run {.run admiral::get_not_mapped()} to access the full list."
         )
       )
-    } else if (nrow(temp_not_mapped) == 0) {
+    } else if (check_not_mapped_type != "none" && !some_not_mapped) {
       cli_inform(
         "All {.var {vars2chr(by_vars_left)}} are mapped."
       )